feat: add module keyboard
