+++ index.html
@@ -8,6 +8,6 @@
   <title>RSS Virtual Keyboard</title>
 </head>
 <body>
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
